+++ src/openalea/core/plugin/formatting/util.py
@@ -1,0 +1,20 @@
+# -*- python -*-
+#
+#       OpenAleaLab
+#
+#       Copyright 2015 INRIA-CIRAD-INRA
+#
+#       File author(s): Guillaume Baty <guillaume.baty@inria.fr>
+#
+#       File contributor(s):
+#
+#       Distributed under the Cecill-C License.
+#       See accompanying file LICENSE.txt or copy at
+#           http://www.cecill.info/licences/Licence_CeCILL-C_V1-en.html
+#
+#       OpenAlea WebSite : http://openalea.github.io
+#
+###############################################################################
+
+
+DEFAULT_ICON = "icons/Crystal_Clear_app_kservices.png"
